Crate encrypted_transfers

This library provides the API needed by the chain, the wallet, and the supporting tools to deal with encrypted amounts.

Modules

proofs
Construction and verification of proofs for encrypted transfer transactions.

types
This module provides common types and constants for encrypted transfers.

Functions

aggregate
Combine two encrypted amounts into one. This is only meaningful if both encrypted amounts are encrypted with the same public key, otherwise the result is meaningless.

decrypt_amount
Decrypt a single amount given the helper table.

encrypt_amount_with_fixed_randomness
Make an encryption of a single amount using a fixed randomness.

make_sec_to_pub_transfer_data
Produce the payload of an secret to public amount transaction.

make_transfer_data
Produce the payload of an encrypted amount transaction.

verify_sec_to_pub_transfer_data
Verify a secret to public amount transaction.

verify_transfer_data
Verify an encrypted amount transaction.