Module encrypted_transfers::proofs

Construction and verification of proofs for encrypted transfer transactions.

Structs

ElgDec

An auxiliary structure that contains data related to the proof of correct decryption. This is stated as an independent protocol in the blue papers, but for efficiency reasons the prover and verifier are inlined together with the main EncTrans proof provers and verifiers. Only the data is separate.

EncTrans

The EncTrans sigma protocol as specified in the blue paper.

EncTransCommit

A structure that represents the intermediate state of the sigma protocol after the prover has committed to all the values they wish to prove statements about. This is then used in the computation of the challenge.

EncTransSecret

Secret values which the EncTrans proof talks about. For constructing proofs these must match the public values that are part of the EncTrans structure.

EncTransState

As for the witness, we don’t need the state for elg_dec

EncTransWitness

Witness for the EncTrans protocol.

Enums

VerificationError

The verifier does three checks. In case verification fails, it can be useful to know which of the checks led to failure.

Functions

gen_enc_trans

Implementation of genEncTrans in the bluepaper

gen_enc_trans_proof_info

Implementation of genEncTransProofInfo in the bluepaper. It produces a sigma protocol of type EncTrans (see enc_trans.rs)

gen_sec_to_pub_trans

Implementation of genSecToPubTrans in the bluepaper

verify_enc_trans

This function is for verifying that an encrypted transfer has been done corretly. The arguments are

verify_sec_to_pub_trans

This function is for verifying that an encrypted transfer has been done corretly. The arguments are