Function encrypted_transfers::proofs::gen_enc_trans

pub fn gen_enc_trans<C: Curve, R: Rng>(
    context: &GlobalContext<C>, 
    ro: &mut RandomOracle, 
    pk_sender: &PublicKey<C>, 
    sk_sender: &SecretKey<C>, 
    pk_receiver: &PublicKey<C>, 
    index: EncryptedAmountAggIndex, 
    S: &Cipher<C>, 
    s: Amount, 
    a: Amount, 
    csprng: &mut R
) -> Option<EncryptedAmountTransferData<C>>

Implementation of genEncTrans in the bluepaper

This function produces transfer data containing a proof that an encrypted transfer was done correctly The arguments are

global context with parameters for generating proofs, and generators for encrypting amounts.

a random oracle needed for the sigma protocol and Bulletproofs

public key and secret key of sender

public key of receiver

index indicating which amounts where used

S - encryption of the input amount up to the index, combined into one encryption

s - input amount

a - amount to send The proof contained in the transfer data produced by this function is a combination a proof produced by the EncTrans sigma protocol and a rangeproof (Bulletproofs), i.e. additonally showing that all a_j and s_j’ are in [0, 2^chunk_size) It returns None if s < a or if it fails to produce one of the bulletproofs.

This implementation differs from the bluepaper in the following ways:

The challenge (ctx in the paper) differs. In the paper this function produces the challenge, but here it is assumed that a random oracle to be used by the sigma protocol and bulletproof is supplied in the correct state. This function is called by encrypted_transfers/src/lib.rs by make_transfer_data where the random oracle provided is in the following state: Domain separator “EncryptedTransfer”, appended with append_message(b“ctx“, global_context), then append_message(b“receiver_pk“, receiver_pk), then append_message(b“sender_pk“, sender_pk)

The generators for the bulletproofs are provided as input through the context: GlobalContext parameter. The rest of the information needed for the bulletproof are the randomness returned by gen_enc_trans_proof_info

The returned value is not signed, we only return the data to be signed by the sender