Function encrypted_transfers::proofs::gen_enc_trans_proof_info

pub fn gen_enc_trans_proof_info<C: Curve>(
    pk_sender: &PublicKey<C>, 
    pk_receiver: &PublicKey<C>, 
    S: &Cipher<C>, 
    A: &[Cipher<C>], 
    S_prime: &[Cipher<C>], 
    h: &C
) -> EncTrans<C>

Implementation of genEncTransProofInfo in the bluepaper. It produces a sigma protocol of type EncTrans (see enc_trans.rs)

Here, both A and S_prime are encrypted amounts that are encrypted in chunks in the exponent, i.e. A is of the form (A_1, …, A_t) where A_i = (g^r_i, h^a_i pk_receiver^r_i) =: (c_{i,1}, c_{i,2}), and where a_i is the i’th chunk of the amount that A is an encryption of. Similarly, S_prime of the form (S_1’, …, S_(t’)’), where S_i’ = (g^r_i’, h^s_i’ pk_sender^r_i’) =: (d_{i,1}, d_{i,2})

This implementation differs from the one defined in the Cryptoprim Bluepaper in the following way:

It takes h, the base for encryption in the exponent, as an input.

We don’t produce the Bulletproof information. This is computed independently

Instead of using the genAndComp, genEqComp and genLinRelCompEx to compose the sigmaprotocol as in the paper, we immediately output EncTrans{zeta_1, zeta_2, zeta_3, zeta_4} and guarantee through the implementation of EncTrans the equality of the decryption key in the dlog and elg-dec protocol, and the linear relation between the chunks of S’, S and A.

See EncTrans for more detail