Function encrypted_transfers::proofs::gen_sec_to_pub_trans

pub fn gen_sec_to_pub_trans<C: Curve, R: Rng>(
    context: &GlobalContext<C>, 
    ro: &mut RandomOracle, 
    pk: &PublicKey<C>, 
    sk: &SecretKey<C>, 
    index: EncryptedAmountAggIndex, 
    S: &Cipher<C>, 
    s: Amount, 
    a: Amount, 
    csprng: &mut R
) -> Option<SecToPubAmountTransferData<C>>

Implementation of genSecToPubTrans in the bluepaper

For sending secret balance to public balance. This function produces transfer data containing a proof that a secret to public balance transfer was done correctly. The arguments are

global context with parameters for generating proofs, and generators for encrypting amounts.

a random oracle needed for the sigma protocol and Bulletproofs

public key and secret key of sender (who is also the receiver)

index indicating which amounts where used

S - encryption of the input amount up to the index, combined into one encryption

s - input amount

a - amount to send

The proof contained in the transfer data produced by this function is a combination a proof produced by the EncTrans sigma protocol and a rangeproof (Bulletproofs), i.e. additonally showing that all s_j’ are in [0, 2^chunk_size). Here, the A that is given to gen_enc_trans_proof_info is an encryption of a with randomness 0, in one chunk. The produced EncTrans is then used to prove that s = a + \sum_{j=1}^(t’) 2^{(chunk_size)*(j-1)} s_j’, where the s_j’ denote the chunks of s’ := s-a. It returns None if s < a, if it fails to produce the sigma proof or if it fails to produce the bulletproofs.

This implementation differs from the bluepapers in the following ways: The challenge (ctx in the paper) differs. In the paper this function produces the challenge, but here it is assumed that a random oracle to be used by the sigma protocol and bulletproof is supplied in the correct state. This function is called by encrypted_transfers/src/lib.rs by make_sec_to_pub_transfer_data where the random oracle provided is in the following state: Domain separator “SecToPubTransfer”, appended with append_message(b“ctx“, global_context), then append_message(b“pk“, pk);

In the bluepaper, a seperate function genSecToPubProofInfo is used to produce the information needed to prove correctness of the transaction. In this implementation, we instead reuse the genEncTransProofInfo function by making a trivial encryption A of the amount to send with randomness = 0 under the public key 1, that is A = (0, h^a). The protocol given by genSecToPubProofInfo in the bluepaper provides a protocol for proving

Knowledge of decryption key of the sender account

Knowledge of (s, sk) such that the secret amount decrypts to s under sk.

The two decryption keys in 1 and 2 are equal

Knowledge of (s’,r) such that S’ (the encrypted remaining amount) is an encryption of s’

Proof of the linear relation that S’ is an encryption of the value s-a, i.e. the value encrypted by S and the amount a to send

All of this is also proved by using genEncTransProofInfo and can be verified since the verifier can produce the same encryption A from a.