Function encrypted_transfers::proofs::verify_enc_trans

pub fn verify_enc_trans<C: Curve>(
    context: &GlobalContext<C>, 
    ro: &mut RandomOracle, 
    transaction: &EncryptedAmountTransferData<C>, 
    pk_sender: &PublicKey<C>, 
    pk_receiver: &PublicKey<C>, 
    S: &Cipher<C>
) -> Result<(), VerificationError>

This function is for verifying that an encrypted transfer has been done corretly. The arguments are

global context with parameters for generating proofs, and generators for encrypting amounts.

a random oracle needed for the sigma protocol and Bulletproofs

a transaction containing a proof

public keys of both sender and receiver

S - Encryption of amount on account It either returns Ok() indicating that the transfer has been done correctly or a VerificationError indicating what failed (the EncTrans protocol or one of the bulletproofs)

This function is only responsible of for checking the cryptographic proofs of an encrypted transfer. This means it varies from the bluepaper in the following way:

In the bluepaper, this function is responsible for checking that the sender account has no aggregatable secret amount, however in the implementation this responsibility is handled by aggregating all amounts at indices less than the index of the transaction.

In the bluepaper, this function is also responsible for checking the signature of the transaction, however this is done elsewhere in the implementation, namely before these cryptographic proofs are checked.