Module encrypted_transfers::types

This module provides common types and constants for encrypted transfers.

Structs

AggregatedDecryptedAmount

An aggregated encrypted amount with a decrypted plaintext, collecting encrypted amounts with decryption. The only real difference from the above is the meaning of the index field.

EncryptedAmount

An encrypted amount, in two chunks in “little endian limbs”. That is, the first chunk represents the low 32 bits of an amount, and the second chunk represents the high 32 bits. The JSON serialization of this is just base16 encoded serialized chunks.

EncryptedAmountAggIndex

An index that represents which encrypted amounts have been combined into an associated encrypted amount. This is in contrast to EncryptedAmountIndex which identifies a single encrypted amount (per account).

EncryptedAmountIndex

A sequential index of an incoming encrypted amount on an account.

EncryptedAmountRandomness

Randomness used when producing an encrypted amount.

EncryptedAmountTransferData

Data that will go onto an encrypted amount transfer.

EncryptedAmountTransferProof

Proof that an encrypted transfer data is well-formed

IndexedEncryptedAmount

An encrypted amount that we know the index of.

SecToPubAmountTransferData

Data that will go onto a secret to public amount transfer.

SecToPubAmountTransferProof

Proof that an encrypted transfer data is well-formed

Constants

CHUNK_SIZE

Size of the chunk for encrypted amounts.